+++ middleman-core/lib/middleman-core/application.rb
@@ -256,10 +256,17 @@ module Middleman
 
     def evaluate_configuration
       # Check for and evaluate local configuration in `config.rb`
-      local_config = File.join(root, 'config.rb')
-      if File.exist? local_config
-        logger.debug '== Reading: Local config'
-        config_context.instance_eval File.read(local_config), local_config, 1
+      config_rb = File.join(root, 'config.rb')
+      if File.exist? config_rb
+        logger.debug '== Reading: Local config: config.rb'
+        config_context.instance_eval File.read(config_rb), config_rb, 1
+      else
+        # Check for and evaluate local configuration in `middleman.rb`
+        middleman_rb = File.join(root, 'middleman.rb')
+        if File.exist? middleman_rb
+          logger.debug '== Reading: Local middleman: middleman.rb'
+          config_context.instance_eval File.read(middleman_rb), middleman_rb, 1
+        end
       end
 
       env_config = File.join(root, 'environments', "#{config[:environment]}.rb")